换电新工程

+++ app/src/main/java/com/ruimeng/things/home/bean/CountAmountBean.kt
@@ -7,8 +7,8 @@ class CountAmountBean (
 
         ){
     data class Data(
-        var totalPrice: Int = 0,
-        var couponAmount: Int = 0
+        var totalPrice: Double = 0.0,
+        var couponAmount: Double = 0.0
     )
 
 }
